+++ src/main/java/net/soulsweaponry/client/SoulsWeaponryClient.java
@@ -4,13 +4,10 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandlerRegistry;
 import net.fabricmc.fabric.api.client.render.fluid.v1.SimpleFluidRenderHandler;
-import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.render.*;
 import net.minecraft.client.render.block.entity.EndPortalBlockEntityRenderer;
 import net.minecraft.util.Identifier;
 import net.soulsweaponry.SoulsWeaponry;
-import net.soulsweaponry.client.hud.CustomBossBar;
-import net.soulsweaponry.client.hud.PostureHudOverlay;
 import net.soulsweaponry.client.registry.*;
 import net.soulsweaponry.networking.PacketRegistry;
 import net.soulsweaponry.registry.BlockRegistry;
@@ -55,15 +52,13 @@ public class SoulsWeaponryClient implements ClientModInitializer {
         BlockRenderLayerMap.INSTANCE.putBlocks(RenderLayer.getTranslucent(), BlockRegistry.VERGLAS_BLOCK);
         BlockRenderLayerMap.INSTANCE.putFluids(RenderLayer.getTranslucent(), FluidRegistry.STILL_PURIFIED_BLOOD, FluidRegistry.FLOWING_PURIFIED_BLOOD);
 
-        HudRenderCallback.EVENT.register(new PostureHudOverlay());
-        CustomBossBar.init();
-
         EntityModelLayerModRegistry.initClient();
         EntityModelRegistry.initClient();
         PredicateRegistry.initClient();
         PacketRegistry.registerS2CPackets();
         KeyBindRegistry.initClient();
         ParticleClientRegistry.initClient();
+        HudRegistry.init();
 
         FluidRenderHandlerRegistry.INSTANCE.register(FluidRegistry.STILL_PURIFIED_BLOOD, FluidRegistry.FLOWING_PURIFIED_BLOOD,
                 new SimpleFluidRenderHandler(