+++ src/main/java/net/soulsweaponry/items/sword/Skofnung.java
@@ -16,6 +16,8 @@ import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
+import net.soulsweaponry.entitydata.BleedData;
+import net.soulsweaponry.entitydata.IEntityDataSaver;
 import net.soulsweaponry.items.ModdedSword;
 import net.soulsweaponry.items.SkofnungStone;
 import net.soulsweaponry.registry.EffectRegistry;
@@ -45,7 +47,8 @@ public class Skofnung extends ModdedSword {
         int duration = (int) (ConfigConstructor.skofnung_disable_heal_duration + (WeaponUtil.getEnchantDamageBonus(stack) * 40));
         target.addStatusEffect(new StatusEffectInstance(EffectRegistry.DISABLE_HEAL, duration, 0));
         if (isEmpowered(stack)) {
-            target.addStatusEffect(new StatusEffectInstance(EffectRegistry.BLEED, 80, 0));
+            BleedData.addBleed((IEntityDataSaver) target, (int) ConfigConstructor.skofnung_empowered_bleed_post_hit);
+            target.addStatusEffect(new StatusEffectInstance(EffectRegistry.BLEED, (int) ConfigConstructor.skofnung_empowered_bleed_effect_duration, (int) ConfigConstructor.skofnung_empowered_bleed_effect_amp));
             if (attacker instanceof PlayerEntity player) {
                 if (!player.getItemCooldownManager().isCoolingDown(this)) {
                     this.reduceEmpowered(stack, player.getWorld(), attacker);