+++ src/main/java/net/soulsweaponry/networking/PacketRegistry.java
@@ -33,5 +33,6 @@ public class PacketRegistry {
         ClientPlayNetworking.registerGlobalReceiver(PacketIds.SYNC_UMBRAL_DAMAGE_COOLDOWN, UTDamageCooldownSyncS2C::receive);
         ClientPlayNetworking.registerGlobalReceiver(PacketIds.STOP_BOSS_MUSIC, StopBossMusicS2C::receive);
         ClientPlayNetworking.registerGlobalReceiver(PacketIds.CHAIN_LIGHTNING, ChainLightningS2C::receive);
+        ClientPlayNetworking.registerGlobalReceiver(PacketIds.BLEED_SYNC, BleedSyncS2C::receive);
     }
 }