+++ src/main/java/net/soulsweaponry/registry/EffectRegistry.java
@@ -1,6 +1,5 @@
 package net.soulsweaponry.registry;
 
-import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.effect.StatusEffect;
@@ -13,17 +12,14 @@ import net.minecraft.potion.Potions;
 import net.minecraft.recipe.BrewingRecipeRegistry;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
-import net.minecraft.registry.tag.EntityTypeTags;
 import net.minecraft.util.Identifier;
 import net.soulsweaponry.SoulsWeaponry;
 import net.soulsweaponry.entity.effect.*;
-import net.soulsweaponry.util.CustomDamageSource;
-import net.soulsweaponry.util.ModTags;
 
 public class EffectRegistry {
 
     public static final StatusEffect HALLOWED_DRAGON_MIST = registerEffect(new HallowedDragonMist(), "hallowed_dragon_mist");
-    public static final StatusEffect BLOODTHIRSTY = registerEffect(new DefaultStatusEffect(StatusEffectCategory.NEUTRAL, 0x630109), "bloodthirsty");
+    public static final StatusEffect BLOODTHIRSTY = registerEffect(new Bloodthirsty(), "bloodthirsty");
     public static final StatusEffect POSTURE_BREAK = registerEffect(new PostureBreak(), "posture_break");
     public static final StatusEffect LIFE_LEACH = registerEffect(new LifeLeach(), "life_leach");
     public static final StatusEffect RETRIBUTION = registerEffect(new DefaultStatusEffect(StatusEffectCategory.HARMFUL, 0xc76700), "retribution");
@@ -33,7 +29,7 @@ public class EffectRegistry {
     public static final StatusEffect MOON_HERALD = registerEffect(new DefaultStatusEffect(StatusEffectCategory.BENEFICIAL, 0x03e8fc), "moon_herald");
     public static final StatusEffect FREEZING = registerEffect(new Freezing(), "freezing");
     public static final StatusEffect DISABLE_HEAL = registerEffect(new DefaultStatusEffect(StatusEffectCategory.HARMFUL, 0xfc9d9d), "disable_heal");
-    public static final StatusEffect BLEED = registerEffect(new DefaultStatusEffect(StatusEffectCategory.HARMFUL, 0xba0c00), "bleed");
+    public static final StatusEffect BLEED = registerEffect(new Bleed(), "bleed");
     public static final StatusEffect CALCULATED_FALL = registerEffect(new DefaultStatusEffect(StatusEffectCategory.BENEFICIAL, 0xffffff), "calculated_fall");
     public static final StatusEffect VEIL_OF_FIRE = registerEffect(new VeilOfFire(), "veil_of_fire");
     public static final StatusEffect BLIGHT = registerEffect(new DefaultStatusEffect(StatusEffectCategory.HARMFUL, 0x73013c), "blight");
@@ -79,36 +75,6 @@ public class EffectRegistry {
 
         public DefaultStatusEffect(StatusEffectCategory statusEffectCategory, int color) {
             super(statusEffectCategory, color);
-        }
-
-        @Override
-        public boolean canApplyUpdateEffect(int duration, int amplifier) {
-            if (this == EffectRegistry.BLEED) {
-                int i = 15 >> amplifier;
-                if (i > 0) {
-                    return duration % i == 0;
-                }
-                return true;
-            }
-            if (this == EffectRegistry.BLOODTHIRSTY) {
-                int k = 40 >> amplifier;
-                if (k > 0) {
-                    return duration % k == 0;
-                } else {
-                    return true;
-                }
-            }
-            return false;
-        }
-        
-        @Override
-        public void applyUpdateEffect(LivingEntity entity, int amplifier) {
-            if (this == EffectRegistry.BLEED && !entity.getType().isIn(EntityTypeTags.SKELETONS) && !entity.getType().isIn(ModTags.Entities.SKELETONS)) {
-                entity.damage(CustomDamageSource.create(entity.getWorld(), CustomDamageSource.BLEED), 1f + amplifier);
-            }
-            if (this == EffectRegistry.BLOODTHIRSTY) {
-                entity.damage(entity.getWorld().getDamageSources().wither(), 1f);
-            }
         }
     }
 }